Add 4.21*: Blog List Expansion, step 9

+++ controllers/blogs.js
@@ -27,8 +27,16 @@ blogsRouter.post('/', async (request, response, next) => {
 
 })
 blogsRouter.delete('/:id', async(request, response)=>{
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const blog = await Blog.findById(request.params.id)
+  if(decodedToken.id === blog.user.toString()){
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  }
+  else return response.status(401).json({ error: 'Unauthorized user' })
 })
 blogsRouter.put('/:id', async(request, response)=>{
   const body = request.body
